Use a simpler sphinx theme (material) with no custom CSS.
This is better to make sure the docs can be edited easily by
many others. Also we need to privilege something with a white
bg so that some pictures look nice. I wanted to use the
readthedocs thene, but it does not wrap lines in tables, and we
have a strong need for this.